requirements.txt
falcon
numpy
scikit-learn
mysql-connector-python-rf
mecab-python3
gensim
